refactor: Update communication settings URL and APP_ID
Update the URL and APP_ID in the communication settings seed file to use the production environment instead of the local development environment.

Remove unused code in the add-coredatasource seed file

Remove the code related to the "Household by Caste" and "Vulnerability Status" data cards in the add-coredatasource seed file.

Add new user and user role in the seed.user.ts file

Add a new user named "Raghav" with the email "raghav.kattel@rumsan.net" and a wallet address to the seed.user.ts file. Also, assign a new user role to this user.

Refactor the main function in the seed.user.ts file

Refactor the main function in the seed.user.ts file to use a transaction for creating roles, permissions, users, auths, and user roles.